requirements.txt
# Core Dependencies
streamlit==1.31.0
python-dotenv==1.0.0

# LLM and AI
openai==1.12.0
anthropic==0.18.1
langchain==0.1.9
langchain-openai==0.0.6
langchain-anthropic==0.1.4
langchain-community==0.0.24

# Data Processing
pandas==2.2.0
numpy==1.26.3
pydantic==2.6.1

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# Report Generation
reportlab==4.0.9
matplotlib==3.8.2
plotly==5.18.0

# Utilities
requests==2.31.0
tenacity==8.2.3
tiktoken==0.6.0

# Development
pytest==8.0.0
black==24.1.1
flake8==7.0.0
